new auth command
