+++ build-tools/src/commands/mod.rs
@@ -1,3 +1,3 @@
 pub mod build_command;
 pub mod deploy_command;
-
+pub mod auth_command;
