+++ build-tools/src/main.rs
@@ -1,4 +1,4 @@
-use commands::{build_command::build, deploy_command::deploy};
+use commands::{build_command::build, deploy_command::deploy, auth_command::auth};
 use structopt::StructOpt;
 
 use std::io::Result;
@@ -9,6 +9,8 @@ mod commands;
 enum SubCommand {
     #[structopt(name = "build", about = "Build extension")]
     Build(commands::build_command::BuildOptions),
+    #[structopt(name = "auth", about = "Auth to AWS")]
+    Auth(commands::auth_command::AuthOptions),
     #[structopt(name = "deploy", about = "Deploy to AWS")]
     Deploy(commands::deploy_command::DeployOptions)
 }
@@ -24,6 +26,9 @@ struct CLI {
 async fn main() -> Result<()> {
     let args = CLI::from_args();
     match args.cmd {
+        SubCommand::Auth(opt) => {
+            auth(&opt).await
+        }
         SubCommand::Build(opt) => {
             build(&opt)
         }
